Updated HFD Mini App
